requirements.txt
openpyxl==3.1.2
pandas==2.0.3
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0